# myProject/las/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "视频社交平台API",
    "version": "1.0.0",
    "description": "视频社交平台的REST API文档",
    "contact": {
      "name": "后端开发团队",
      "email": "backend@example.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/messages": {
      "get": {
        "summary": "获取所有留言",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "页码"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "每页数量"
          }
        ]
      }
    },
    "/messages/create": {
      "post": {
        "summary": "创建新留言",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/messages/{id}": {
      "get": {
        "summary": "获取留言详情",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "put": {
        "summary": "更新留言",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "delete": {
        "summary": "删除留言",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/messages/{id}/like": {
      "post": {
        "summary": "点赞留言",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/messages/{id}/comments": {
      "get": {
        "summary": "获取留言的评论",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "summary": "注册新用户",
        "tags": [
          "认证"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户注册成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "用户登录",
        "tags": [
          "认证"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "刷新访问令牌",
        "tags": [
          "认证"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "令牌刷新成功"
          },
          "401": {
            "description": "刷新令牌无效或已过期"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "获取当前用户信息",
        "tags": [
          "用户"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "用户信息获取成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "patch": {
        "summary": "更新当前用户信息",
        "tags": [
          "用户"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户信息更新成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "获取所有用户列表（管理员）",
        "tags": [
          "用户"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "页码（默认1）"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "每页数量（默认10）"
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    },
    "/users/{userId}/role": {
      "patch": {
        "summary": "更改用户角色（管理员）",
        "tags": [
          "用户"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "admin"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户角色更改成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "认证",
      "description": "用户认证相关API"
    },
    {
      "name": "用户",
      "description": "用户管理相关API"
    }
  ]
}